#647 Add and display option arrivalDelay
* Show arrival time if arrivalDelay is set

TODO: It would be nice to-have if the time until the arrival time updates dynamically
Currently it is calculated once.

Also we should take into account if the phone's clock is wrong.

For this we could fetch the server time during plan fetch from
plan.date and calculate the offset with the phone clock and use that to correct the time stamp.